pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = "my blog"
authors = ["정명진 <thinkjin99@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.0.7"
flake8 = "^7.1.0"
pytest = "^8.2.2"
pytest-django = "^4.8.0"
django-debug-toolbar = "^4.4.6"
pydantic = "^2.8.2"
pyjwt = "^2.8.0"
cryptography = "^42.0.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
